Centralize TEI detection logic
